refactor(file upload): move logic to components dir
